# Back/RecaudosAPI/RecaudosAPI/Models/EstacionDetalle.cs
namespace RecaudosAPI.Models
{
    public class EstacionDetalle
    {
        public int Id { get; set; }
        public string? cantidad { get; set; }
        public string? valor { get; set; }

    }
}