# Back/RecaudosAPI/RecaudosAPI/Models/Fechas.cs
using System.Security.Policy;

namespace RecaudosAPI.Models
{
    public class Fechas
    {
        public int Id { get; set; }
        public DateTime fecha { get; set; }
        public string? totalCantidad { get; set; }
        public string? totalValor { get; set; }
        public string? estacion { get; set; }

    }
}