# Documentos/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RecaudosAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Recaudos": {
      "get": {
        "tags": [
          "Recaudos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recaudos"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recaudos"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recaudos"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recaudos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recaudos"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Recaudos"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Recaudos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Recaudos"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recaudos"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recaudos"
                }
              }
            }
          }
        }
      }
    },
    "/api/Recaudos/GetTotales": {
      "get": {
        "tags": [
          "Recaudos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Totales"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Totales"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Totales"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Recaudos/GetFechas": {
      "get": {
        "tags": [
          "Recaudos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fechas"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fechas"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fechas"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Recaudos/GetEstaciones": {
      "get": {
        "tags": [
          "Recaudos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Estaciones"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Estaciones"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Estaciones"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Recaudos/{id}": {
      "get": {
        "tags": [
          "Recaudos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Recaudos"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recaudos"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recaudos"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Recaudos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recaudos"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Recaudos"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Recaudos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Recaudos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EstacionDetalle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "string",
            "nullable": true
          },
          "valor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Estaciones": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "estacion": {
            "type": "string",
            "nullable": true
          },
          "totalCantidadF": {
            "type": "integer",
            "format": "int32"
          },
          "totalValorF": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "estacionDetalle": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EstacionDetalle"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Fechas": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "totalCantidad": {
            "type": "string",
            "nullable": true
          },
          "totalValor": {
            "type": "string",
            "nullable": true
          },
          "estacion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Recaudos": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "estacion": {
            "type": "string",
            "nullable": true
          },
          "sentido": {
            "type": "string",
            "nullable": true
          },
          "hora": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "valorTabulado": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Totales": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "nullable": true
          },
          "totalCantidad": {
            "type": "string",
            "nullable": true
          },
          "totalValor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}